Initiialized D3D11 Backend
Simple Initialization of DX11, creating a device, swapchain and device context for use within the program - in addition to clearing the graphics client region to a colour. 
Minor memory management issues in need of refactoring; I'm considering using WRL::ComPtr<> for this, but would like to have some practice with explicitly calling Release().