Fog FX + Resource Allocator
Added one of my Effects to support Fog and simple lighting - I might work more on Materials before adding any more, so I have an easy interface for effect variables. At the moment, they're a little awkward to deal with. 
Also dramatically reduced application memory overheads by using a resource manager (as of yet un-threaded, but that's a stone's throw away...)
Before this commit, up to 1GB of memory was being used for resources. Now, that's down to ~128mb. And with further work, I can probably reduce that even further.